+++ week5/mawakeb/src/main/java/next/controller/users/ListUserController.java
@@ -1,7 +1,6 @@
-package next.controller;
+package next.controller.users;
 
 import core.db.DataBase;
-import next.model.User;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -11,19 +10,20 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/users/profile")
-public class ProfileController extends HttpServlet {
+@WebServlet("/users")
+public class ListUserController extends HttpServlet {
     private static final long serialVersionUID = 1L;
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String userId = req.getParameter("userId");
-        User user = DataBase.findUserById(userId);
-        if (user == null) {
-            throw new NullPointerException("사용자를 찾을 수 없습니다.");
+        if (!UserSessionUtils.isLogined(req.getSession())) {
+            resp.sendRedirect("/users/loginForm");
+            return;
         }
-        req.setAttribute("user", user);
-        RequestDispatcher rd = req.getRequestDispatcher("/user/profile.jsp");
+
+        req.setAttribute("users", DataBase.findAll());
+
+        RequestDispatcher rd = req.getRequestDispatcher("/user/list.jsp");
         rd.forward(req, resp);
     }
 }
