+++ week5/mawakeb/src/main/java/next/controller/users/ProfileController.java
@@ -1,6 +1,7 @@
-package next.controller;
+package next.controller.users;
 
 import core.db.DataBase;
+import next.model.User;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -10,20 +11,19 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/users")
-public class ListUserController extends HttpServlet {
+@WebServlet("/users/profile")
+public class ProfileController extends HttpServlet {
     private static final long serialVersionUID = 1L;
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        if (!UserSessionUtils.isLogined(req.getSession())) {
-            resp.sendRedirect("/users/loginForm");
-            return;
+        String userId = req.getParameter("userId");
+        User user = DataBase.findUserById(userId);
+        if (user == null) {
+            throw new NullPointerException("사용자를 찾을 수 없습니다.");
         }
-
-        req.setAttribute("users", DataBase.findAll());
-
-        RequestDispatcher rd = req.getRequestDispatcher("/user/list.jsp");
+        req.setAttribute("user", user);
+        RequestDispatcher rd = req.getRequestDispatcher("/user/profile.jsp");
         rd.forward(req, resp);
     }
 }
