+++ week5/mawakeb/src/main/java/next/controller/users/UpdateUserController.java
@@ -1,4 +1,4 @@
-package next.controller;
+package next.controller.users;
 
 import java.io.IOException;
