+++ week5/mawakeb/src/main/java/next/controller/users/UserSessionUtils.java
@@ -1,4 +1,4 @@
-package next.controller;
+package next.controller.users;
 
 import javax.servlet.http.HttpSession;
